don't use mlockall by default
is_InitCamera hung because our RLIMIT_MEMLOCK ulimit was too low (64B)
which meant that once that much memory was locked from future
allocations, mmap would begin to fail with EAGAIN. This also means you
cannot allocate stack space for a thread you would want to host with
mmap. Since mmap with MAP_STACK is used internally, thread creation
failed entirely, which meant that the parent was left waiting on a
value indefinitely as the thread meant to wake it up was never
created, therefore deadlock!