port to a proper state machine
This makes state management centralized, eliminates global bool fatal,
and paves the way for managing multiple Camera instances in the same
driver instance at once. We also have a set of fixed states the camera
object can be in, and this mechanism is extendable (i.e. addition of
sub-states in the future for more conditional code).